ai_workflow_agent/requirements.txt
boto3==1.35.27 #The AWS SDK for Python
fastapi==0.115.0 #FastAPI framework, high performance, easy to learn, fast to code, ready for production
fastapi-cli==0.0.5 #Run and manage FastAPI apps from the command line with FastAPI CLI
langchain==0.3.1 #Building applications with LLMs through composability
langgraph==0.2.28 #Building stateful, multi-actor applications with LLMs
langchain_aws==0.2.1 #An integration package connecting AWS and LangChain
langchain-community==0.3.1 #Community contributed LangChain integrations
langchain-core==0.3.6 #Building applications with LLMs through composability
langchain_anthropic==0.2.1 #An integration package connecting AnthropicMessages and LangChain
pydantic==2.9.2 #Data validation using Python type hints
pydantic_core==2.23.4 #Core functionality for Pydantic validation and serialization
pymongo==4.9.1 #Python driver for MongoDB <http://www.mongodb.org>
requests==2.32.3 #Python HTTP for Humans.
streamlit==1.38.0 #A faster way to build and share data apps
transformers==4.45.0 #State-of-the-art Machine Learning for JAX, PyTorch and TensorFlow
uvicorn==0.30.6 #The lightning-fast ASGI server.
faiss-cpu==1.8.0.post1 #A library for efficient similarity search and clustering of dense vectors.
twilio==9.3.2 #Twilio API client and TwiML generator
notion_client==2.2.1 #Python client for the official Notion API
reportlab==4.2.2 #The Reportlab Toolkit
pypdf==5.0.0 #A pure-python PDF library capable of splitting, merging, cropping, and transforming PDF files
python-docx==1.1.2 #Create, read, and update Microsoft Word .docx files
cryptography==43.0.1 #FOR SOME PDF DOCUMENTS IN ADMIN
python-jose==3.3.0 #JOSE implementation in Python
ipython
langgraph-checkpoint-postgres
psycopg-binary==3.2.3
psycopg-pool==3.2.4